Supplier Bankruptcy Risk Assessment and Modeling
In this problem, we would like to assess the bankruptcy risk of potential suppliers. For that purpose, we have collected data related to past instances of bankruptcy of suppliers. The data consists of data related to bankrupt as well as solvent suppliers for comparison purposes. The objective of the project is to build a risk assessment model for potential bankruptcy of suppliers. Bankruptcy of suppliers is a key source of supply chain disruption. A primary purpose of supply chain planning is to assess disruption risk and plan for alternatives.